requirements.txt
# Core dependencies
crewai>=0.1.0
python-dotenv>=0.19.0

# Data handling and analysis
yfinance>=0.2.36
pandas>=2.0.0

# Web tools and APIs
beautifulsoup4>=4.12.0
requests>=2.31.0

# Report generation
markdown2>=2.4.10
pyyaml>=6.0.1

# Optional but recommended
ipython>=8.0.0  # For better REPL experience
jupyter>=1.0.0  # For notebook support 